refactor the Home file structure
